// client/src/models/brewery.ts
export interface Brewery {
    id?: number;
    name?: string;
    brewery_type?: string;
    street?: string;
    city?: string;
    state?: string;
    postal_code?: string;
    country?: string;
    longitude?: number;
    latitude?: number;
    phone?: string;
    website_url?: string;
    updated_at?: string;
    distance?: number;
}